requirements.txt
snowflake-connector-python
grpcio==1.67.1
python-dotenv
sqlalchemy
snowflake-sqlalchemy
pandas
streamlit
groq
streamlit-aggrid
sqlglot
fuzzywuzzy
snowflake-connector-python[pandas]
pyarrow
python-Levenshtein
